Move constraint code to its own package
This code was formerly part of the scheduler package. It needs to be
moved to an independent package so the orchestrator can use it when
considering where to create tasks for global services.

While moving it, made a few minor cleanups:

- Renamed Expr to Constraint
- Unexport Key field
- Renamed ParseExprs to Parse
