+++ manager/constraint/constraint_test.go
@@ -1,4 +1,4 @@
-package scheduler
+package constraint
 
 import (
 	"testing"
@@ -6,91 +6,91 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestParseExprs(t *testing.T) {
+func TestParse(t *testing.T) {
 	// empty string
-	_, err := ParseExprs([]string{""})
+	_, err := Parse([]string{""})
 	assert.Error(t, err)
 
-	_, err = ParseExprs([]string{" "})
+	_, err = Parse([]string{" "})
 	assert.Error(t, err)
 
 	// no operator
-	_, err = ParseExprs([]string{"nodeabc"})
+	_, err = Parse([]string{"nodeabc"})
 	assert.Error(t, err)
 
 	// incorrect operator
-	_, err = ParseExprs([]string{"node ~ abc"})
+	_, err = Parse([]string{"node ~ abc"})
 	assert.Error(t, err)
 
-	// Cannot use the leading digit for Key
-	_, err = ParseExprs([]string{"1node==a2"})
+	// Cannot use the leading digit for key
+	_, err = Parse([]string{"1node==a2"})
 	assert.Error(t, err)
 
 	// leading and trailing white space are ignored
-	_, err = ParseExprs([]string{" node == node1"})
+	_, err = Parse([]string{" node == node1"})
 	assert.NoError(t, err)
 
-	// Key cannot container white space in the middle
-	_, err = ParseExprs([]string{"no de== node1"})
+	// key cannot container white space in the middle
+	_, err = Parse([]string{"no de== node1"})
 	assert.Error(t, err)
 
-	// Cannot use * in Key
-	_, err = ParseExprs([]string{"no*de==node1"})
+	// Cannot use * in key
+	_, err = Parse([]string{"no*de==node1"})
 	assert.Error(t, err)
 
-	// Key cannot be empty
-	_, err = ParseExprs([]string{"==node1"})
+	// key cannot be empty
+	_, err = Parse([]string{"==node1"})
 	assert.Error(t, err)
 
 	// value cannot be empty
-	_, err = ParseExprs([]string{"node=="})
+	_, err = Parse([]string{"node=="})
 	assert.Error(t, err)
 
 	// value cannot be an empty space
-	_, err = ParseExprs([]string{"node== "})
+	_, err = Parse([]string{"node== "})
 	assert.Error(t, err)
 
-	// Cannot use $ in Key
-	_, err = ParseExprs([]string{"no$de==node1"})
+	// Cannot use $ in key
+	_, err = Parse([]string{"no$de==node1"})
 	assert.Error(t, err)
 
-	// Allow CAPS in Key
-	exprs, err := ParseExprs([]string{"NoDe==node1"})
+	// Allow CAPS in key
+	exprs, err := Parse([]string{"NoDe==node1"})
 	assert.NoError(t, err)
-	assert.Equal(t, exprs[0].Key, "NoDe")
+	assert.Equal(t, exprs[0].key, "NoDe")
 
-	// Allow dot in Key
-	exprs, err = ParseExprs([]string{"no.de==node1"})
+	// Allow dot in key
+	exprs, err = Parse([]string{"no.de==node1"})
 	assert.NoError(t, err)
-	assert.Equal(t, exprs[0].Key, "no.de")
+	assert.Equal(t, exprs[0].key, "no.de")
 
 	// Allow leading underscore
-	exprs, err = ParseExprs([]string{"_node==_node1"})
+	exprs, err = Parse([]string{"_node==_node1"})
 	assert.NoError(t, err)
-	assert.Equal(t, exprs[0].Key, "_node")
+	assert.Equal(t, exprs[0].key, "_node")
 
 	// Allow special characters in exp
-	exprs, err = ParseExprs([]string{"node==[a-b]+c*(n|b)/"})
+	exprs, err = Parse([]string{"node==[a-b]+c*(n|b)/"})
 	assert.NoError(t, err)
-	assert.Equal(t, exprs[0].Key, "node")
+	assert.Equal(t, exprs[0].key, "node")
 	assert.Equal(t, exprs[0].exp, "[a-b]+c*(n|b)/")
 
 	// Allow space in Exp
-	exprs, err = ParseExprs([]string{"node==node 1"})
+	exprs, err = Parse([]string{"node==node 1"})
 	assert.NoError(t, err)
-	assert.Equal(t, exprs[0].Key, "node")
+	assert.Equal(t, exprs[0].key, "node")
 	assert.Equal(t, exprs[0].exp, "node 1")
 }
 
 func TestMatch(t *testing.T) {
-	exprs, err := ParseExprs([]string{"node.name==foo"})
+	exprs, err := Parse([]string{"node.name==foo"})
 	assert.NoError(t, err)
 	e := exprs[0]
 	assert.True(t, e.Match("foo"))
 	assert.False(t, e.Match("fo"))
 	assert.False(t, e.Match("fooE"))
 
-	exprs, err = ParseExprs([]string{"node.name!=foo"})
+	exprs, err = Parse([]string{"node.name!=foo"})
 	assert.NoError(t, err)
 	e = exprs[0]
 	assert.False(t, e.Match("foo"))
@@ -98,7 +98,7 @@ func TestMatch(t *testing.T) {
 	assert.True(t, e.Match("fo"))
 	assert.True(t, e.Match("fooExtra"))
 
-	exprs, err = ParseExprs([]string{"node.name==f*o"})
+	exprs, err = Parse([]string{"node.name==f*o"})
 	assert.NoError(t, err)
 	e = exprs[0]
 	assert.False(t, e.Match("fo"))
@@ -109,7 +109,7 @@ func TestMatch(t *testing.T) {
 	assert.False(t, e.Match("foo"))
 
 	// test special characters
-	exprs, err = ParseExprs([]string{"node.name==f.-$o"})
+	exprs, err = Parse([]string{"node.name==f.-$o"})
 	assert.NoError(t, err)
 	e = exprs[0]
 	assert.False(t, e.Match("fa-$o"))
